Adding various files somehow untracked.

+++ coderunner/Sandbox/liusandbox.php
@@ -10,17 +10,62 @@
 
 require_once('sandboxbase.php');
 
-class LiuSandbox extends Sandbox {
-    private $LANGUAGES = array('C', 'python2', 'python3');
-    private $LANG_VERSIONS = array('C' => 'gcc-4.6.3',
-        'python2' => 'Python 2.7.3', 'python3' => 'Python 3.2.3');
 
-    // For each language, either define its compiler or its executable
-    // TODO: generalise the 'compilers' option for non-C -style languages
-    // (currently the -o filename flag is added by the 'compile' method).
-    private $COMPILERS = array('C' => 'gcc -Wall -ansi -lm -x c ');
-    private $EXECUTABLES = array('python2' => '/usr/bin/python2',
-        'python3' => '/usr/bin/python3');
+class LiuSandbox extends Sandbox {
+    private $LANGUAGES = array(
+        'C'         => array(
+            'version'      => 'gcc-4.6.3',
+            'compile'      => 'gcc -Wall -ansi -static -lm -x c -o {{SOURCE_FILE}}.exe {{SOURCE_FILE}} 2>{{SOURCE_FILE}}.err',
+            'executable'   => '{{SOURCE_FILE}}.exe',
+            'run'          => '{{COPIED_EXECUTABLE}}',
+            'readableDirs' => array()
+        ),
+
+        'python2'   => array(
+            'version'      => 'Python 2.7.3',
+            'compile'      => '',
+            'executable'   => '{{SOURCE_FILE}}',
+            'run'          => '/usr/bin/python2 -BESs {{COPIED_EXECUTABLE}}',
+            'readableDirs' => array(
+                                '/lib/',
+                                '/lib64/',
+                                '/etc/',
+                                '/usr/local/lib',
+                                '/usr/lib',
+                                '/usr/bin',
+                                '/proc/meminfo',
+                                '/usr/include'
+                            )
+        ),
+
+        'python3'   => array(
+            'version'      => 'Python 3.2.3',
+            'compile'      => '',
+            'executable'   => '{{SOURCE_FILE}}',
+            'run'          => '/usr/bin/python3 -BESs {{COPIED_EXECUTABLE}}',
+            'readableDirs' => array(
+                                '/lib/',
+                                '/lib64/',
+                                '/opt/python3',
+                                '/etc/',
+                                '/usr/local/lib',
+                                '/usr/lib',
+                                '/usr/bin',
+                                '/proc/meminfo',
+                                '/usr/include'
+                            )
+        ),
+
+        // Matlab isn't a go-er in the Liu sandbox. Not sure what will happen here.
+        'matlab'    => array(
+            'version'      => 'Matlab R2012a',
+            'compile'      => 'cp {{SOURCE_FILE}} {{SOURCE_FILE}}.m  2>{{SOURCE_FILE}}.err',
+            'executable'   => '{{SOURCE_FILE}}.m',
+            'run'          => "/usr/local/Matlab2012a/bin/glnxa64/MATLAB -nodisplay -nojvm -r {{SOURCE_FILE_BASENAME}}()",
+            'readableDirs' => array('/')    // **** TODO **** FIX ME
+        )
+    );
+
 
     private $RESULT_CODES = array(
             'PD' => Sandbox::RESULT_INTERNAL_ERR,  // Pending???
@@ -46,7 +91,8 @@ class LiuSandbox extends Sandbox {
     }
 
     public function getLanguages() {
-        return (object) array('error' => Sandbox::OK, 'languages' => $this->LANGUAGES);
+        return (object) array('error' => Sandbox::OK,
+            'languages' => array_keys($this->LANGUAGES));
     }
 
 
@@ -66,47 +112,55 @@ class LiuSandbox extends Sandbox {
      * @return object with 'error' field (always '') and 'link' field (always $currentRunId above)
      * @throws coding_exception if I've goofed
      *
-     * TODO: add caching of most-recently-compiled C program (or a few such).
      */
     public function createSubmission($sourceCode, $language, $input,
                                         $run=TRUE, $private=TRUE) {
-        if (!in_array($language, $this->LANGUAGES, TRUE)) {
+        if (!isset($this->LANGUAGES[$language])) {
             throw new coding_exception('LiuSandbox::createSubmission: Bad language');
         }
 
-        // Copy the sourceCode into a temporary file
-        $srcfname = tempnam("/tmp", "coderunner_src_");
-        $handle = fopen($srcfname, "w");
-        fwrite($handle, $sourceCode);
-        fclose($handle);
+        $languageDetails = $this->LANGUAGES[$language];
 
         // Record input data in $this in case requested in call to getSubmissionDetails
         $this->source = $sourceCode;
         $this->input = $input;
         $this->language = $language;
+
+        // Copy the sourceCode into a temporary file
+        $srcfname = tempnam("/tmp", "coderunner_");
+        $handle = fopen($srcfname, "w");
+        fwrite($handle, $sourceCode);
+        fclose($handle);
+
         if (!$run || !$private) {
-            throw new conding_exception('LiuSandbox::createSubmission: unexpected param value');
+            throw new coding_exception('LiuSandbox::createSubmission: unexpected param value');
         }
         $this->date = date("Y-m-d H-i-s");
 
+        $errors = '';
+        $errorFileName = $srcfname . '.err';
+        $executableFileName = $this->expand($languageDetails['executable'], $srcfname);
 
-        if (isset($this->COMPILERS[$language])) {
+        if ($languageDetails['compile'] !== '') {
             // Compile if this language requires compilation
             // Optimise (e.g. when running the same program with multiple different
             // stdin data) to avoid recompiling.
 
             if ($this->lastCompiledCode === $sourceCode) {
-                $executable = $this->lastCompileOutputFile;
+                $executableFileName = $this->lastCompileOutputFile;
                 $errors = $this->lastCompileErrors;
             }
             else {
-                list($executable, $errors) = $this->compile($srcfname, $language);
+                $compileCommand = $this->expand($languageDetails['compile'], $srcfname);
+                $errors = $this->compile($compileCommand, $errorFileName);
+
+                // Cache results to avoid recompiling the same code again
+                $this->lastCompiledCode = $this->source;
+                $this->lastCompileErrors = $errors;
+                $this->lastCompileOutputFile = $executableFileName;
             }
 
-            if ($errors === '') {
-                $args = array();
-                $this->runInSandbox($args, $executable, $input);
-            } else {
+            if ($errors !== '') {
                 $this->result = Sandbox::RESULT_COMPILATION_ERROR;
                 $this->cmpinfo = $errors;
                 $this->time = 0;
@@ -115,14 +169,13 @@ class LiuSandbox extends Sandbox {
                 $this->output = '';
                 $this->stderr = '';
             }
-
-        } else {
-            // Interpreted language
-            $errors = '';
-            $args = array($this->EXECUTABLES[$language], '-BESsu');
-            $this->runInSandbox($args, $srcfname, $input);
         }
 
+        if ($errors === '') {
+            $executeCommand = $this->expand($languageDetails['run'], $srcfname);
+            $readableDirs = $languageDetails['readableDirs'];
+            $this->runInSandbox($executeCommand, $executableFileName, $input, $readableDirs);
+        }
 
         unlink($srcfname);
         return (object) array('error' => Sandbox::OK, 'link' => self::$currentRunId);
@@ -150,7 +203,7 @@ class LiuSandbox extends Sandbox {
                             'result'    => $this->result,
                             'langId'    => array_search($this->language, $this->LANGUAGES),
                             'langName'  => $this->language,
-                            'langVersion' => $this->LANG_VERSIONS[$this->language],
+                            'langVersion' => $this->LANGUAGES[$this->language]['version'],
                             'time'      => $this->time,
                             'date'      => $this->date,
                             'memory'    => $this->memory,
@@ -182,47 +235,43 @@ class LiuSandbox extends Sandbox {
      * The caller should already have checked that the source code
      * being compiled isn't exactly the same as last time.
      */
-    private function compile($srcfile, $lang) {
-        if ($this->lastCompileOutputFile != '') {
+    private function compile($compileCommand, $errorFileName) {
+        if ($this->lastCompileOutputFile !== '' &&
+                file_exists($this->lastCompileOutputFile)) {
             unlink($this->lastCompileOutputFile);
         }
-        $exec = $srcfile . '.exe';
-        $errs = $srcfile . '.err';
-        $cmd = $this->COMPILERS[$lang] . " -o $exec $srcfile 2>$errs";
-        exec($cmd, $output, $returnVar);
+        exec($compileCommand, $output, $returnVar);
         if ($returnVar == 0) {
             $compileErrors = '';
         }
         else {
-            $compileErrors = file_get_contents($errs);
-            $exec = NULL;
+            $compileErrors = file_get_contents($errorFileName);
         }
-        unlink($errs);
+        unlink($errorFileName);
 
-        // Cache results to avoid recompiling the same code again
-        $this->lastCompiledCode = $this->source;
-        $this->lastCompileErrors = $compileErrors;
-        $this->lastCompileOutputFile = $exec;
-
-        return array($exec, $compileErrors);
+        return $compileErrors;
     }
 
     // On close, delete the last compiler output file (if we have one).
     public function close() {
-        if ($this->lastCompileOutputFile != '') {
+        if ($this->lastCompileOutputFile !== '' &&
+                file_exists($this->lastCompileOutputFile)) {
             unlink($this->lastCompileOutputFile);
         }
     }
 
 
-    // Run the given command in the sandbox, with the given file and stdin.
+    // Run the given command in the sandbox, after copying the specified
+    // file into the sandbox temporary working directory and using the
+    // given input data as stdin. Read access will be permitted to all
+    // files nominated in $readableDirs.
     // Results are all left in $this for later access by getSubmissionDetails
-    private function runInSandbox($args, $filename, $input) {
+    private function runInSandbox($executeCommand, $filename, $input, $readableDirs) {
        global $CFG;
        // Set up the control params for the sandbox
-
+        //debugging($executeCommand);
         $run = array(
-            'args'      => $args,
+            'cmd'       => $executeCommand,
             'filename'  => $filename,
             'input'     => $input,
             'quota'     => array(
@@ -230,7 +279,8 @@ class LiuSandbox extends Sandbox {
                 'cpu'       => 5000,     // 5 secs
                 'memory'    => 64000000, // 64MB
                 'disk'      => 1048576   // 1 MB
-            )
+            ),
+            'readableDirs' => $readableDirs
         );
 
         // Write the control params to a file, JSON encoded,
@@ -247,6 +297,7 @@ class LiuSandbox extends Sandbox {
 
         $cmd = $CFG->dirroot . "/question/type/coderunner/Sandbox/liusandbox.py $tmpfname";
         exec($cmd, $output, $returnVar);
+
         $outputJson = $output[0];
         $result = json_decode($outputJson);
 
@@ -260,13 +311,21 @@ class LiuSandbox extends Sandbox {
             $this->signal = $result->details->signal_info[0];
             $this->time = $result->details->elapsed;
             $this->memory = $result->details->mem_info[0];
-        } else {  // Python2 probe in Liu sandbox doesn't return some of the info
-            $this->signal = print_r($result->details, TRUE);
+        } else {  // An internal error doesn't set all fields
+            $this->signal = 0;
             $this->time = 0;
             $this->memory = 0;
         }
 
         unlink($tmpfname);
+    }
+
+
+    // Expand a given template, replacing all occurrences of '{{SOURCE_FILE}}'
+    // with the given actual file name. This is really a hook for more elaborate
+    // templating later if needed.
+    private function expand($template, $filename) {
+        return str_replace('{{SOURCE_FILE}}', $filename, $template);
     }
 
 }
